requirements.txt
# Core Trading & Data
python-binance==1.0.19
vectorbt==0.26.2
pandas==2.2.0
numpy==1.26.3
ccxt==4.2.25

# Optimization & ML
optuna==3.5.0
scikit-learn==1.4.0
river==0.21.0

# Async & Scheduling
aiohttp==3.9.3
websockets==12.0
APScheduler==3.10.4

# Monitoring & Observability
prometheus-client==0.19.0
python-telegram-bot==20.7
python-json-logger==2.0.7

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Code Quality
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
mypy==1.8.0
ruff==0.1.14
black==24.1.1
pre-commit==3.6.0

# Type Stubs
types-requests==2.31.0.20240125
pandas-stubs==2.1.4.231227
